refactor: context 적용 개수 증가에 따른 리펙토링
